dodanie klasy EmailConfirmation ( wysylanie maila potwierdzajacego dokonanie platnosci)
